reafctor: split gql queries to separate files
