+++ src/constants/query/getCategories.js
@@ -1,0 +1,9 @@
+import { gql } from '@apollo/client';
+
+export const GET_CATEGORIES = gql`
+  query GetCategories  {
+    categories {
+      name
+    }
+  }
+`;
