+++ src/constants/query/getCurrencies.js
@@ -1,0 +1,10 @@
+import { gql } from '@apollo/client';
+
+export const GET_CURRENCIES = gql`
+  query GetCurrencies {
+    currencies {
+      label
+      symbol
+    }
+  }
+`;
